requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
supabase==2.0.2
google-generativeai==0.3.2
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
hypothesis==6.92.1
pytest==7.4.3
pytest-django==4.7.0
twilio==8.10.3
reportlab==4.0.7
Pillow==10.1.0
weasyprint==61.2
pydyf<0.11.0
qrcode[pil]==7.4.2